Move instruments to its own folder
