+++ Assets/00_Content/Scripts/Interactables/Instruments.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 231c8ab3fd6d4f05a5dfa9ba9db87d97
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
