+++ Assets/00_Content/Scripts/Interactables/Instruments/InstrumentData.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Interactables {
+namespace Interactables.Instruments {
 	[CreateAssetMenu(fileName = "Instrument", menuName = "Game/Instrument", order = 0)]
 	public class InstrumentData : ScriptableObject {
 		public AudioClip music;
